requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.1
django-allauth==0.50.0
django-cors-headers==3.11.0
django-extensions==3.1.5
django-import-export==2.7.1
djangorestframework==3.13.1
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.3
MarkupPy==1.14
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.9
Pillow==9.0.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
razorpay==1.3.0
requests==2.27.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
tablib==3.2.0
tzdata==2021.5
urllib3==1.26.8
whitenoise==6.0.0
xlrd==2.0.1
xlwt==1.3.0
